Separate independent JS components ( SOFTWARE-5097 )
