+++ src/static/js/components/wrapper.js
@@ -1,0 +1,12 @@
+import {remove_children} from "../util.js";
+
+class Wrapper {
+    constructor() {
+        this.node = document.getElementById("wrapper")
+    }
+    remove_children = () => {
+        remove_children(this.node)
+    }
+}
+
+export {Wrapper}
