+++ src/static/js/pages/resources.js
@@ -1,5 +1,5 @@
 import {lod_to_ul} from "../util.js";
-import {TablePage} from "../table.js"
+import {TablePage} from "../components/table.js"
 
 const data_function = async () => {
     let data = await fetch("/miscresource/json").then(d => d.json())
